further revisions
