+++ qwandaq/src/main/java/life/genny/qwandaq/utils/QwandaUtils.java
@@ -839,64 +839,6 @@ public class QwandaUtils {
 
 		return ask;
 	}
-	
-	/**
-	 * Get the edit question groups for a {@link BaseEntity}
-	 * <p> will default to <b>QUE_BASEENTITY_GRP</b> if no {@link Attribute#LNK_EDIT_PCMS} attribute exists in the {@link Definition}
-	 *     or if the value is blank
-	 * </p>
-	 * @param baseEntityCode a BaseEntity or Definition code 
-	 * @return array of question group codes in order of appearance for editing the BaseEntity
-	 */
-	public String getEditPcmCodes(String targetCode) {
-		if(targetCode == null)
-			throw new NullParameterException("targetCode");
-		
-		// ensure def
-		log.info("[!] Attempting to retrieve edit pcms from base entity: " + targetCode);
-		Definition baseEntity = defUtils.getDEF(targetCode);
-		if(baseEntity == null) {
-			log.error("Could not find Definition of be: " + targetCode);
-			throw new BadDataException("No definition of BaseEntity: " + targetCode);
-		}
-
-		return getEditPcmCodes(baseEntity);
-	}
-
-	/**
-	 * Get the edit question groups for a {@link BaseEntity}
-	 * <p> will default to <b>QUE_BASEENTITY_GRP</b> if no {@link Attribute#LNK_EDIT_PCMS} attribute exists in the {@link Definition}
-	 *     or if the value is blank
-	 * </p>
-	 * @param baseEntityCode a BaseEntity or Definition code 
-	 * @return array of question group codes in order of appearance for editing the BaseEntity
-	 */
-	public String getEditPcmCodes(BaseEntity baseEntity) {
-		if(baseEntity == null)
-			throw new NullParameterException("baseEntity");
-		
-		// Convert to DEF
-		if(!baseEntity.getCode().startsWith(Prefix.DEF_)) {
-			baseEntity = defUtils.getDEF(baseEntity);
-		}
-
-		// Look for attached edit ques. If none then default to QUE_BASEENTITY_GRP
-		Optional<EntityAttribute> editQuesLnk = baseEntity.findEntityAttribute(Attribute.LNK_EDIT_PCMS);
-		if(!editQuesLnk.isPresent()) {
-			log.warn("Could not find LNK_EDIT_PCMS in " + baseEntity.getCode() + ". Defaulting to PCM_FORM_EDIT");
-			return "[PCM_FORM_EDIT]";
-		}
-		
-		String editQues = editQuesLnk.get().getValueString();
-		log.debug("FOUND LNK_EDIT_PCMS in BaseEntity: " + baseEntity + ". Edit PCM Codes: " + editQues);
-
-		if(!StringUtils.isBlank(editQues)) {
-			return editQues;
-		}
-		
-		log.warn("Edit pcm was blank. Defaulting to PCM_FORM_EDIT");
-		return "[PCM_FORM_EDIT]";
-	}
 
 	/**
 	 * Update the status of the disabled field for an Ask on the web.
